more newflow feature helper cleanup - signup feature tests working
