--- spec/cassettes/Newflow/Students/student_signup_flow/sf_setup.yml
@@ -1,56 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://test.salesforce.com/services/oauth2/token
-    body:
-      encoding: US-ASCII
-      string: grant_type=password&client_id=<salesforce_consumer_key>&client_secret=<salesforce_consumer_secret>&username=<salesforce_username_url>&password=<salesforce_password_url><salesforce_security_token>
-    headers:
-      User-Agent:
-      - Faraday v1.0.1
-      Content-Type:
-      - application/x-www-form-urlencoded
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Tue, 29 Mar 2022 18:37:28 GMT
-      Set-Cookie:
-      - BrowserId=SX-pOq-PEeycW30MxhCiLg; domain=.salesforce.com; path=/; expires=Wed,
-        29-Mar-2023 18:37:28 GMT; Max-Age=31536000
-      - CookieConsentPolicy=0:0; path=/; expires=Wed, 29-Mar-2023 18:37:28 GMT; Max-Age=31536000
-      - LSKey-c$CookieConsentPolicy=0:0; path=/; expires=Wed, 29-Mar-2023 18:37:28
-        GMT; Max-Age=31536000
-      Strict-Transport-Security:
-      - max-age=63072000; includeSubDomains
-      X-Content-Type-Options:
-      - nosniff
-      X-Xss-Protection:
-      - 1; mode=block
-      Cache-Control:
-      - no-cache,must-revalidate,max-age=0,no-store,private
-      Expires:
-      - Thu, 01 Jan 1970 00:00:00 GMT
-      X-Readonlymode:
-      - 'false'
-      Content-Type:
-      - application/json;charset=UTF-8
-      Vary:
-      - Accept-Encoding
-      Transfer-Encoding:
-      - chunked
-      Cf-Team:
-      - 1078910dac00009f0001935400000001
-    body:
-      encoding: ASCII-8BIT
-      string: '{"access_token":"<salesforce_access_token>","instance_url":"<salesforce_instance_url>","id":"<salesforce_id>","token_type":"Bearer","issued_at":"1648579049383","signature":"<salesforce_signature>"}'
-    http_version:
-  recorded_at: Tue, 29 Mar 2022 18:37:29 GMT
-recorded_with: VCR 3.0.3
