+++ spec/support/feature_helpers.rb
@@ -241,6 +241,21 @@ def strip_html(text)
   ActionView::Base.full_sanitizer.sanitize(text)
 end
 
+def expect_sheerid_iframe
+  within_frame do
+    expect(page).to have_text('Verify your instructor status')
+    expect(page.find('#sid-country')[:value]).to have_text('United States', exact: false)
+    expect(page.find('#sid-teacher-school')[:value]).to be_blank
+    expect(page.find('#sid-first-name')[:value]).to have_text(first_name)
+    expect(page.find('#sid-last-name')[:value]).to have_text(last_name)
+    expect(page.find('#sid-email')[:value]).to have_text(email_value)
+    expect(page).to have_text('Can\'t find your country in the list? Click here.')
+    expect(page).to have_text('Can\'t find your school in the list? Click here.')
+    expect(page).to have_text('Verify my instructor status')
+    screenshot!
+  end
+end
+
 def expect_security_log(*args)
   expect_any_instance_of(ActionController::Base).to receive(:security_log)
                                                       .with(*args)
